Fix: set dtype to int in cyclic permutation function

+++ qudit-bb/utils.py
@@ -6,6 +6,6 @@ def cyclic_permutation(dim : int, shift : int) -> np.ndarray:
     """Construct cyclic shift permutation matrix of size dim x dim, shifted by shift."""
 
     if shift == 0:
-        return np.identity(dim)
+        return np.identity(dim, dtype=int)
     else:
-        return np.roll(np.identity(dim), shift, axis=1)
+        return np.roll(np.identity(dim, dtype=int), shift, axis=1)
